Move file system policy unit tests to the separate file
Otherwise importing pytest module kills the import.
